% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-archive.R
\name{bin_archive}
\alias{bin_archive}
\title{Get QR code for bin on Filebin}
\usage{
bin_archive(bin, format = "zip", file = NA)
}
\arguments{
\item{bin}{Bin name. If \code{NA} the generate random bin name.}

\item{format}{Archive format. Either \code{"zip"} (default) or \code{"tar"}.}
}
\value{

}
\description{
Get QR code for bin on Filebin
}
\examples{
bin_archive("placeholder")
}
