% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-delete.R
\name{bin_delete}
\alias{bin_delete}
\title{Delete a bin from Filebin}
\usage{
bin_delete(bin)
}
\arguments{
\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A Boolean indicating whether bin was successfully deleted.
}
\description{
Delete a bin from Filebin
}
\examples{
\dontrun{
bin_delete("placeholder")
bin_delete("https://filebin.net/placeholder/")
}
}
