% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-get.R
\name{bin_get}
\alias{bin_get}
\title{Retrieve a bin from Filebin}
\usage{
bin_get(bin)
}
\arguments{
\item{filename}{File name or URL.}

\item{file}{File name to use to save results. If \code{FALSE} then don't save
to file.}

\item{overwrite}{Whether to overwrite existing file.}
}
\value{

}
\description{
Retrieve a bin from Filebin
}
\examples{
bin_get("placeholder")
}
