% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-put.R
\name{bin_lock}
\alias{bin_lock}
\title{Lock a bin on Filebin}
\usage{
bin_lock(bin)
}
\arguments{
\item{filename}{File name or URL.}

\item{file}{File name to use to save results. If \code{FALSE} then don't save
to file.}

\item{overwrite}{Whether to overwrite existing file.}
}
\value{

}
\description{
Lock a bin on Filebin
}
\examples{
bin_get("placeholder")
}
