% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bin_name_random}
\alias{bin_name_random}
\title{Create random bin name for Filebin}
\usage{
bin_name_random(length = 16, pattern = "[a-z0-9]")
}
\arguments{
\item{length}{Number of characters in bin name. Must be 8 or more.}

\item{pattern}{Character set.}
}
\value{
A character vector.
}
\description{
Create random bin name for Filebin
}
\examples{
bin_name_random(8)
}
