% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-qr-code.R
\name{bin_qr_code}
\alias{bin_qr_code}
\title{Get QR code for bin on Filebin}
\usage{
bin_qr_code(bin, file = NA)
}
\arguments{
\item{bin}{Bin name. If \code{NA} the generate random bin name.}

\item{file}{File name to use to save results.}
}
\value{
File name of the archive.
}
\description{
Get QR code for bin on Filebin
}
\examples{
\dontrun{
posted <- file_post(LOREM_IPSUM, bin = "latin-text")
bin_qr_code("latin-text", file = tempfile(fileext = ".png"))
}
}
