% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-delete.R
\name{file_delete}
\alias{file_delete}
\title{Delete a file from Filebin}
\usage{
file_delete(filename, bin)
}
\arguments{
\item{filename}{File name or URL.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A Boolean indicating whether file was successfully deleted.
}
\description{
Delete a file from Filebin
}
\examples{
\dontrun{
posted <- file_post(LOREM_IPSUM, bin = "latin-text")

file_delete("lorem-ipsum.txt", "latin-text")
file_delete("https://filebin.net/latin-text/lorem-ipsum.txt")
}
}
