% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-get.R
\name{file_get}
\alias{file_get}
\title{Retrieve a file from Filebin}
\usage{
file_get(filename, bin, file = NA, overwrite = FALSE)
}
\arguments{
\item{filename}{File name or URL.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}

\item{file}{File name to use to save results. If \code{FALSE} then don't save
to file.}

\item{overwrite}{Whether to overwrite existing file.}
}
\value{
Contents of file.
}
\description{
Retrieve a file from Filebin
}
\examples{
#'
library(stringi)

lorem_ipsum <- tempfile(fileext = ".txt")
bin <- bin_name_random()

cat(stri_rand_lipsum(3), sep = "\n\n", file = lorem_ipsum)
file_post(lorem_ipsum, bin)

# Discard path and just retain filename.
filename <- basename(lorem_ipsum)
file_get(filename, bin)

# Delete downloaded file.
file.remove(filename)
}
