% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-post.R
\name{file_post}
\alias{file_post}
\title{Upload a file to Filebin}
\usage{
file_post(path, bin = NA)
}
\arguments{
\item{path}{Path to a file. If a vector then upload multiple files.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A tibble.
}
\description{
Upload a file to Filebin
}
\examples{
# Create some sample data.
library(stringi)

lorem_ipsum <- tempfile(fileext = ".txt")
ipsum_lorem <- tempfile(fileext = ".txt")

cat(stri_rand_lipsum(3), sep = "\n\n", file = lorem_ipsum)
cat(stri_rand_lipsum(7), sep = "\n\n", file = ipsum_lorem)

# Upload a single file.
file_post(lorem_ipsum)
# Upload multiple files.
file_post(c(lorem_ipsum, ipsum_lorem))

# Upload to a specific bin.
bin <- bin_name_random()
file_post(lorem_ipsum, bin)

# To get details of the HTTP request.
httr::with_verbose(file_post(lorem_ipsum))
}
