% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-get.R
\name{filebin_get}
\alias{filebin_get}
\title{Retrieve a file from Filebin}
\usage{
filebin_get(filename, bin, file = NA, overwrite = FALSE)
}
\arguments{
\item{filename}{File name or URL.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}

\item{file}{File name to use to save results. If \code{FALSE} then don't save
to file.}

\item{overwrite}{Whether to overwrite existing file.}
}
\value{

}
\description{
Retrieve a file from Filebin
}
\examples{
filebin_get("lorem-ipsum.txt", "placeholder")
filebin_get("https://filebin.net/placeholder/lorem-ipsum.txt")
}
