% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-post.R
\name{filebin_post}
\alias{filebin_post}
\title{Upload a file to Filebin}
\usage{
filebin_post(path, bin = NA)
}
\arguments{
\item{path}{Path to a file. If a vector then upload multiple files.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A tibble.
}
\description{
Upload a file to Filebin
}
\examples{
# Create some sample data.
library(stringi)

cat(stri_rand_lipsum(3), sep = "\n\n", file = "lorem-ipsum.txt")
cat(stri_rand_lipsum(7), sep = "\n\n", file = "ipsum-lorem.txt")

# Upload a single file.
filebin_post("lorem-ipsum.txt")
# Upload multiple files.
filebin_post(c("lorem-ipsum.txt", "ipsum-lorem.txt"))
# Upload to a specific bin.
filebin_post("lorem-ipsum.txt", "placeholder")

# To get details of the HTTP request.
httr::with_verbose(filebin_post("lorem-ipsum.txt"))
}
