% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-post.R
\name{post}
\alias{post}
\title{Upload a file to a bin}
\usage{
post(path, bin = NA)
}
\arguments{
\item{path}{Path to a file. If a vector then upload multiple files.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A tibble.
}
\description{
Upload a file to a bin
}
\examples{
# Upload a single file.
post("/etc/passwd")
# Upload multiple files.
post(c("/etc/passwd", "/etc/group"))
# Upload to a specific bin.
post("/etc/passwd", "examples")

# To get details of the HTTP request.
httr::with_verbose(post("/etc/passwd"))
}
