% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem-ipsum.R
\docType{data}
\name{sample-data-path}
\alias{sample-data-path}
\alias{LOREM_IPSUM}
\alias{MORE_LOREM_IPSUM}
\title{Paths of sample data files}
\format{
An object of class \code{NULL} of length 0.

An object of class \code{NULL} of length 0.
}
\usage{
LOREM_IPSUM

MORE_LOREM_IPSUM
}
\description{
Paths of sample data files
}
\keyword{datasets}
