% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-get.R
\name{bin_get}
\alias{bin_get}
\title{Retrieve a bin from Filebin}
\usage{
bin_get(bin)
}
\arguments{
\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A tibble.
}
\description{
Retrieve a bin from Filebin
}
\examples{
\dontrun{
posted <- file_post(LOREM_IPSUM)
bin_get(posted$bin)
}
}
