% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-post.R
\name{file_post}
\alias{file_post}
\title{Upload a file to Filebin}
\usage{
file_post(path, bin = NA)
}
\arguments{
\item{path}{Path to a file. If a vector then upload multiple files.}

\item{bin}{Bin name. If \code{NA} the generate random bin name.}
}
\value{
A tibble.
}
\description{
Upload a file to Filebin
}
\examples{
\dontrun{
# Upload a single file.
file_post(LOREM_IPSUM)
# Upload multiple files.
file_post(c(LOREM_IPSUM, MORE_LOREM_IPSUM))

# Upload to a specific bin.
bin <- bin_name_random(length = 24)
file_post(LOREM_IPSUM, bin)
}
}
